Patch 6.1.403 (extra) Problem:    MS-Windows 16 bit: compiler warnings. Solution:   Add typecasts. (Walter Briscoe) Files:	    src/ex_cmds2.c, src/gui_w48.c, src/os_mswin.c, src/os_win16.c, 	    src/syntax.c

+++ src/gui_w48.c
@@ -2595,7 +2595,7 @@ mch_set_mouse_shape(int shape)
 # ifdef WIN32
 	SetClassLong(s_textArea, GCL_HCURSOR, (LONG)LoadCursor(NULL, idc));
 # else
-	SetClassWord(s_textArea, GCW_HCURSOR, LoadCursor(NULL, idc));
+	SetClassWord(s_textArea, GCW_HCURSOR, (WORD)LoadCursor(NULL, idc));
 # endif
 #endif
 	if (!p_mh)
